Remove Guava hack dependency from UserdevConfigs
In a Minecraft context we expect to have Guava, also we use Gradle with its richer module metadata model. So we don't need this dependency. Additionally, using listenablefuture separate from Guava is an extreme edge case in a Minecraft context, because, as previously mentioned, we expect to have Guava on the classpath. In that case though, the user should use a plugin like `java-ecosystem-capabilities` to declare attributes. This also fixes using `java-ecosystem-capabilities` along with NeoForge.